requirements.txt
python-telegram-bot[job-queue]
google-generativeai
gspread
oauth2client
python-dotenv
flask
gunicorn
requests
pandas
matplotlib
openpyxl
